feat: add custom hook fo header
