+++ src/utils/interfaces.ts
@@ -93,3 +93,20 @@ export interface FilterI {
   };
   filteredCatalog?: ProductSearchResult[];
 }
+
+export type UseMobileMenuResult = {
+  menuOpen: boolean;
+  setMenuOpen: React.Dispatch<React.SetStateAction<boolean>>;
+  closeMenu: () => void;
+  openMenu: () => void;
+  toggleMenu: () => void;
+};
+
+export type AuthSectionProps = {
+  onItemClick?: () => void;
+};
+
+export type NavLinksProps = {
+  onItemClick?: () => void;
+};
+
